#369 Rexport all public functionality of lib in index.js
